+++ CHANGES.md
@@ -5,6 +5,7 @@
 - **ENH: Add a classify function to classify a raster according to bins and values** ([#40](https://github.com/sertit/sertit-utils/issues/40))
 - FIX: Fix `ci.assert_val` for xr.DataArrayCoordinates
 - FIX: Update also `long_name` with `new_name` in `rasters.set_metadata`
+- FIX: Fix rasterizing vectors with multiband rasters ([#43](https://github.com/sertit/sertit-utils/issues/43))
 
 ## 1.46.4 (2025-04-04)